feat: EXLM-272 image decoration
