+++ blocks/img-md/img-md.scss
@@ -1,4 +1,4 @@
-.img.block {
+.img-md.block {
   display: flex;
 
   &.left-align {
